Extract create post
